fix: move pulse generator from helper to own file to reduce memory footprint when imported
